# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transaction statistics",
        "description": "REST Api",
        "contact": {
            "name": "Gianluca Siciliano",
            "email": "gianluca.siciliano.79@gmail.com"
        },
        "version": "1.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "API Server"
        }
    ],
    "paths": {
        "/statistics": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "returns the statistic based of the transactions of the last 60 seconds.",
                "description": "returns the statistic based of the transactions of the last 60 seconds.",
                "operationId": "82f865d2c1bf5c0495acee18a9d5e855",
                "responses": {
                    "200": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "returns all the transaction from a specified timestamp",
                "description": "returns all the transaction from a specified timestamp",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "description": "the given timestamp in the ISO 8601 format",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2021-11-27T18:13:25.354Z"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "This endpoint is called to create a new transaction.",
                "description": "This endpoint is called to create a new transaction.",
                "operationId": "store",
                "requestBody": {
                    "description": "Amount",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/transactionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "204": {
                        "description": "Resource created but is older than 60 seconds",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Transactions"
                ],
                "summary": "This endpoint causes all existing transactions to be deleted",
                "description": "This endpoint causes all existing transactions to be deleted",
                "operationId": "destroy",
                "responses": {
                    "200": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "transactionRequest": {
                "properties": {
                    "amount": {
                        "type": "string",
                        "format": "numeric",
                        "example": "1000"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-11-27T18:13:25.354Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "in": "header",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "http://localhost:8080/oauth/token",
                        "scopes": {}
                    }
                },
                "scheme": [
                    "http",
                    "https"
                ]
            }
        }
    },
    "security": [
        []
    ]
}